backend/api-service/requirements.txt
fastapi
uvicorn>=0.21.0
motor>=3.1.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
pymongo>=4.3.3
python-dateutil>=2.8.2
prometheus-client>=0.16.0
python-dotenv>=1.0.0 